// frontend/src/api/types.gen.ts
export type GameSessionSummary = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string | null;
    /**
     * Create Timestamp
     */
    create_timestamp: string;
    /**
     * Nb Parties
     */
    nb_parties: number;
    /**
     * Scores
     */
    scores: Array<PlayerScoreSummary>;
};
// ...
export type PlayerScoreSummary = {
    /**
     * Player Id
     */
    player_id: number;
    /**
     * Score
     */
    score: number;
    /**
     * Player
     */
    player: string;
};